requirements.txt
azure-core==1.23.1
azure-identity==1.7.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.1
Django==2.2.28
django-extensions==3.1.3
django-simple-history==2.7.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.2
dnspython==2.2.1
idna==3.3
josepy==1.13.0
mailer==0.8.1
mozilla-django-oidc==1.2.2
msal==1.17.0
msal-extensions==0.3.1
msgraph-core==0.2.2
numpy==1.19.5
pandas==1.0.3
portalocker==2.4.0
py-topping==0.3.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.1
python-ldap==3.4.0
pytz==2022.1
pywin32==303
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
xlrd==1.2.0
